# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "API Suas Respostas",
        "description": " API para plataformas que visa permitir que pessoas publiquem questões e respostas de atividades e trabalhos.",
        "version": "1.0",
        "contact": {
            "email": "geoo677@gmail.com"
        }
    },
    "servers": [
        {
            "url": "https://api-suas-questoes.herokuapp.com/",
            "description": "API em produção",
            "security": [
                {
                    "bearerAuth": []
                }
            ]
        },
        {
            "url": "http://localhost:8080/",
            "description": "API localhost",
            "security": [
                {
                    "bearerAuth": []
                }
            ]
        }
    ],
    "basePath": "/",
    "paths": {
...
        "/answers/{questionId}?page={page}": {
            "get": {
                "tags": [
                    "Sugestões de Respostas"
                ],
                "summary": "Listar respostas de uma questão.",
                "description": "Lista as respostas de uma questão.",
                "operationId": "getAnswers",
                "parameters": [
                    {
                        "name": "questionId",
                        "in": "path",
                        "description": "Id da questão que vai listar respostas.",
                        "required": true,
                        "default": "1"
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Paginação.",
                        "required": true,
                        "default": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listas de respostas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Answers"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não existe uma questão com esse id.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao listar respostas.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/answers/list/by/user?page={page}": {
            "get": {
                "tags": [
                    "Sugestões de Respostas"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Listar sugestões de usuário respostas de uma questão.",
                "description": "Lista as sugestões de respostas feita pelo usuário.",
                "operationId": "getAnswersByUser",
                "parameters": [
                    
                    {
                        "name": "page",
                        "in": "path",
                        "description": "Paginação.",
                        "required": true,
                        "default": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listas de sugestões de respostas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Answers"
                                }
                            }
                        }
                    },
                    
                    "500": {
                        "description": "Erro ao listar respostas.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/answers/{questionId}": {
            "post": {
                "tags": [
                    "Sugestões de Respostas"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Sugerir resposta para questão.",
                "description": "Salvar sugestão de resposta para questão.",
                "operationId": "postAnswer",
                "parameters": [
                    {
                        "name": "questionId",
                        "in": "path",
                        "description": "Id da questão que vai responder.",
                        "required": true,
                        "default": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Answer"
                            },
                            "examples": {
                                "answer": {
                                    "value": {
                                        "answer": "Resposta: a) 2"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Confirmação de que a sugestão de resposta foi salva",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Acesso negado!.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não existe uma questão com esse id.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Resposta é necessária!.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao salvar sugestão de resposta.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/answers/{questionId}/{answerId}": {
            "put": {
                "tags": [
                    "Sugestões de Respostas"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Editar sugestão de respostas.",
                "description": "Edita sugestão de resposta feita pelo usuário logado.",
                "operationId": "postAnswer",
                "parameters": [
                    {
                        "name": "questionId",
                        "in": "path",
                        "description": "Id da questão que vai editar a sugestão de resposta.",
                        "required": true,
                        "default": "1"
                    },
                    {
                        "name": "answerId",
                        "in": "path",
                        "description": "Id da sugestão de resposta.",
                        "required": true,
                        "default": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Answer"
                            },
                            "examples": {
                                "answer": {
                                    "value": {
                                        "answer": "Resposta: b) 3"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Confirmação de que a sugestão de resposta foi editada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Acesso negado!.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não existe uma questão com esse id.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Resposta é necessária!.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao editar sugestão de resposta.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Sugestões de Respostas"
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "summary": "Deletar sugestão de resposta.",
                "description": "Deleta sugestão de resposta para questão.",
                "operationId": "postAnswer",
                "parameters": [
                    {
                        "name": "questionId",
                        "in": "path",
                        "description": "Id da questão em que vai deletar a sugestão de resposta.",
                        "required": true,
                        "default": "1"
                    },
                    {
                        "name": "answerId",
                        "in": "path",
                        "description": "Id da sugestão de resposta.",
                        "required": true,
                        "default": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Confirmação de que a sugestão de resposta foi deletada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Acesso negado!.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não existe uma questão com esse id.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao deletar sugestão de resposta.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Message": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "Answer": {
                "type": "object",
                "properties": {
                    "answer": {
                        "type": "string"
                    }
                }
            },
...
            "Answers": {
                "type": "object",
                "properties": {
                    "limit": {
                        "type": "integer"
                    },
                    "page": {
                        "type": "integer"
                    },
                    "count": {
                        "type": "integer"
                    },
                    "offset": {
                        "type": "integer"
                    },
                    "answers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Answer"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}